requirements.txt
anyio==4.10.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pillow==11.3.0
playwright==1.55.0
pyee==13.0.0
sniffio==1.3.1
soupsieve==2.8
typing_extensions==4.15.0
